+++ RELEASE_NOTES.md
@@ -3,6 +3,11 @@ Release Notes
 
 ## 0.1.0-beta-800
 
+#### ToDo
+
+- router
+- cache
+
 #### Breaking changes
 
 - Removed serialization methods from `Giraffe.Common`
@@ -10,6 +15,11 @@ Release Notes
 - Removed the overloards for `BindJsonAsync<'T>`, `BindModelAsync<'T>` and `WriteJsonAsync` which accepted an object of type `JsonSerializerSettings`
 - Removed the `customJson` http handler function
 - In order for Giraffe to be fully functional one has to register the default Giraffe dependencies by calling `services.AddGiraffe()` from your `RegisterServices` method in the startup code.
+- Renamed `ReturnHtmlFileAsync()` to `WriteHtmlFileAsync`
+- Renamed `RenderHtmlAsync()` to `WriteHtmlViewAsync`
+- Renamed `html` http handler to `htmlString`
+- Renamed `renderHtml` http handler to `htmlView`
+- Renamed `setBodyAsString` http handler to `setBodyFromString`
 
 #### New features
 
@@ -21,6 +31,10 @@ Release Notes
 - Added the `DefaultXmlSerializer` class which implements `IXmlSerializer` and gets auto registered when adding the Giraffe dependencies via `services.AddGiraffe()` in your startup code.
 - Added `PATCH` http handler to `Giraffe.TokenRouter`
 - Added `HEAD`, `OPTIONS`, `TRACE`, `CONNECT` http handlers
+- The renamed method `WriteHtmlFileAsync` checks now if the path is rooted or not and makes the right decision to find the file on disk
+- Added `WriteHtmlStringAsync` to `Giraffe.HttpContextExtensions`
+- Added `GetHostingEnvironment` to `Giraffe.HttpContextExtensions`
+- Added `SetContentType` to `Giraffe.HttpContextExtensions`
 
 ## 0.1.0-beta-700